refactored viewmodel, implemented progress bar, added retry button incase network failure, tidying layout
