+++ app/src/main/java/com/babylon/babylonchallenge/common/Constants.kt
@@ -1,0 +1,7 @@
+package com.babylon.babylonchallenge.common
+
+class Constants {
+    companion object{
+        const val BASE_URL = "https://jsonplaceholder.typicode.com"
+    }
+}
