+++ app/src/main/java/com/babylon/babylonchallenge/view/MainActivity.kt
@@ -1,0 +1,23 @@
+package com.babylon.babylonchallenge.view
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.fragment.app.FragmentManager
+import androidx.fragment.app.FragmentTransaction
+import com.babylon.babylonchallenge.R
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+        displayPostFragment()
+    }
+
+    private fun displayPostFragment(){
+        val fragmentManager:FragmentManager = supportFragmentManager
+        val fragmentTransaction:FragmentTransaction = fragmentManager.beginTransaction()
+        fragmentTransaction.add(R.id.fragmentContainer,PostsFragment())
+            .commit()
+    }
+}
