+++ app/src/main/java/com/babylon/babylonchallenge/view/PostAdapter.kt
@@ -1,0 +1,35 @@
+package com.babylon.babylonchallenge.view
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import com.babylon.babylonchallenge.R
+import com.babylon.babylonchallenge.data.model.Post
+
+class PostAdapter(private val listener: PostClickListener) :
+    RecyclerView.Adapter<PostViewHolder>() {
+    private var posts: List<Post> = emptyList()
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PostViewHolder {
+        return PostViewHolder(
+            LayoutInflater.from(parent.context).inflate(
+                R.layout.post_view_holder,
+                parent,
+                false
+            )
+        )
+    }
+
+    override fun getItemCount(): Int {
+        return posts.size
+    }
+
+    override fun onBindViewHolder(holder: PostViewHolder, position: Int) {
+        holder.bind(posts[position], listener)
+    }
+
+    fun setItems(posts: List<Post>) {
+        this@PostAdapter.posts = posts
+        notifyDataSetChanged()
+    }
+
+}
