+++ app/src/main/java/com/babylon/babylonchallenge/view/PostClickListener.kt
@@ -1,0 +1,7 @@
+package com.babylon.babylonchallenge.view
+
+import com.babylon.babylonchallenge.data.model.Post
+
+interface PostClickListener {
+    fun onPostClicked(post: Post)
+}
