--- app/src/main/java/com/example/babylonchallenge/common/Constants.kt
@@ -1,7 +1,0 @@
-package com.example.babylonchallenge.common
-
-class Constants {
-    companion object{
-        const val BASE_URL = "https://jsonplaceholder.typicode.com"
-    }
-}
