--- app/src/main/java/com/example/babylonchallenge/data/model/Comment.kt
@@ -1,13 +1,0 @@
-package com.example.babylonchallenge.data.model.comments
-
-import com.google.gson.annotations.SerializedName
-
-
-data class Comment (
-
-	@SerializedName("postId") val postId : Int,
-	@SerializedName("id") val id : Int,
-	@SerializedName("name") val name : String,
-	@SerializedName("email") val email : String,
-	@SerializedName("body") val body : String
-)
