--- app/src/main/java/com/example/babylonchallenge/data/model/Post.kt
@@ -1,12 +1,0 @@
-package com.example.babylonchallenge.data.model
-import android.os.Parcelable
-import com.google.gson.annotations.SerializedName
-import kotlinx.android.parcel.Parcelize
-
-@Parcelize
-data class Post (
-	@SerializedName("userId") val userId : Int,
-	@SerializedName("id") val id : Int,
-	@SerializedName("title") val title : String,
-	@SerializedName("body") val body : String
-): Parcelable
