--- app/src/main/java/com/example/babylonchallenge/view/PostClickListener.kt
@@ -1,7 +1,0 @@
-package com.example.babylonchallenge.view
-
-import com.example.babylonchallenge.data.model.Post
-
-interface PostClickListener {
-    fun onPostClicked(post: Post)
-}
