+++ app/src/test/java/com/babylon/babylonchallenge/AppTests.kt
@@ -1,0 +1,8 @@
+package com.babylon.babylonchallenge
+
+import org.junit.runner.RunWith
+import org.junit.runners.Suite
+
+@Suite.SuiteClasses(PostDetailViewModelTest::class, PostViewModelTest::class)
+@RunWith(Suite::class)
+class AppTests
