--- app/src/test/java/com/example/babylonchallenge/AppTests.kt
@@ -1,8 +1,0 @@
-package com.example.babylonchallenge
-
-import org.junit.runner.RunWith
-import org.junit.runners.Suite
-
-@Suite.SuiteClasses(PostDetailViewModelTest::class, PostViewModelTest::class)
-@RunWith(Suite::class)
-class AppTests
